+++ NEWS.md
@@ -3,7 +3,7 @@
 - Data frame subsetting is now faster and internally simpler. 
 Data frame negative subsetting is now cheaper.
 
-- New functions `df_slice`, `df_select` and `df_sset` for row slicing and 
+- New functions `sset_df`, `sset_row` and `sset_col` for row slicing and 
 col selecting with minimal overhead (~ 1 microsecond).
 
 # cheapr 1.0.1